+++ tests/Gedmo/Tree/TreeObjectHydratorTest.php
@@ -12,8 +12,6 @@ declare(strict_types=1);
 namespace Gedmo\Tests\Tree;
 
 use Doctrine\Common\EventManager;
-use Doctrine\DBAL\Logging\DebugStack;
-use Doctrine\DBAL\Logging\Middleware;
 use Doctrine\ORM\Query;
 use Gedmo\Tests\Tool\BaseTestCaseORM;
 use Gedmo\Tests\Tree\Fixture\Category;
@@ -49,13 +47,7 @@ final class TreeObjectHydratorTest extends BaseTestCaseORM
         $this->populate();
         $this->em->clear();
 
-        // TODO: Remove the "if" check and "else" body when dropping support of doctrine/dbal 2.
-        if (class_exists(Middleware::class)) {
-            $this->queryLogger->reset();
-        } else {
-            $stack = new DebugStack();
-            $this->em->getConfiguration()->setSQLLogger($stack);
-        }
+        $this->queryLogger->reset();
 
         $repo = $this->em->getRepository(self::ROOT_CATEGORY);
 
@@ -96,12 +88,7 @@ final class TreeObjectHydratorTest extends BaseTestCaseORM
         static::assertCount(0, $citrons->getChildren());
 
         // Make sure only one query was executed
-        // TODO: Remove the "if" check and "else" body when dropping support of doctrine/dbal 2.
-        if (class_exists(Middleware::class)) {
-            static::assertCount(1, $this->queryLogger->queries);
-        } else {
-            static::assertCount(1, $stack->queries);
-        }
+        static::assertCount(1, $this->queryLogger->queries);
     }
 
     public function testPartialTreeHydration(): void
@@ -109,13 +96,7 @@ final class TreeObjectHydratorTest extends BaseTestCaseORM
         $this->populate();
         $this->em->clear();
 
-        // TODO: Remove the "if" check and "else" body when dropping support of doctrine/dbal 2.
-        if (class_exists(Middleware::class)) {
-            $this->queryLogger->reset();
-        } else {
-            $stack = new DebugStack();
-            $this->em->getConfiguration()->setSQLLogger($stack);
-        }
+        $this->queryLogger->reset();
 
         /** @var NestedTreeRepository $repo */
         $repo = $this->em->getRepository(self::ROOT_CATEGORY);
@@ -140,12 +121,7 @@ final class TreeObjectHydratorTest extends BaseTestCaseORM
         static::assertSame('Citrons', $citrons->getTitle());
         static::assertCount(0, $citrons->getChildren());
 
-        // TODO: Remove the "if" check and "else" body when dropping support of doctrine/dbal 2.
-        if (class_exists(Middleware::class)) {
-            static::assertCount(2, $this->queryLogger->queries);
-        } else {
-            static::assertCount(2, $stack->queries);
-        }
+        static::assertCount(2, $this->queryLogger->queries);
     }
 
     public function testMultipleRootNodesTreeHydration(): void
@@ -153,13 +129,7 @@ final class TreeObjectHydratorTest extends BaseTestCaseORM
         $this->populate();
         $this->em->clear();
 
-        // TODO: Remove the "if" check and "else" body when dropping support of doctrine/dbal 2.
-        if (class_exists(Middleware::class)) {
-            $this->queryLogger->reset();
-        } else {
-            $stack = new DebugStack();
-            $this->em->getConfiguration()->setSQLLogger($stack);
-        }
+        $this->queryLogger->reset();
 
         /** @var NestedTreeRepository $repo */
         $repo = $this->em->getRepository(self::ROOT_CATEGORY);
@@ -196,12 +166,7 @@ final class TreeObjectHydratorTest extends BaseTestCaseORM
         static::assertSame('Citrons', $citrons->getTitle());
         static::assertCount(0, $citrons->getChildren());
 
-        // TODO: Remove the "if" check and "else" body when dropping support of doctrine/dbal 2.
-        if (class_exists(Middleware::class)) {
-            static::assertCount(2, $this->queryLogger->queries);
-        } else {
-            static::assertCount(2, $stack->queries);
-        }
+        static::assertCount(2, $this->queryLogger->queries);
     }
 
     protected function getUsedEntityFixtures(): array