+++ CHANGELOG.md
@@ -21,6 +21,7 @@ a release.
 ### Changed
 - Dropped support for PHP < 7.4
 - Dropped support for Symfony < 5.4
+- Dropped support for doctrine/dbal < 3.2
 
 ### Deprecated
 - Calling `Gedmo\Mapping\Event\Adapter\ORM::getObjectManager()` and `getObject()` on EventArgs that do not implement `getObjectManager()` and `getObject()` (such as old EventArgs implementing `getEntityManager()` and `getEntity()`)